+++ src/derived_systems/stroboscopic_map.jl
@@ -15,7 +15,7 @@ The second signature first creates a [`CoupledODEs`](@ref) and then calls the fi
 `StroboscopicMap` follows the [`DynamicalSystem`](@ref) interface.
 In addition, the function `set_period!(smap, period)` is provided,
 that sets the period of the system to a new value (as if it was a parameter).
-As this system is in discrete time, [`current_time`](@Ref) and [`initial_time`](@ref)
+As this system is in discrete time, [`current_time`](@ref) and [`initial_time`](@ref)
 are integers. The initial time is always 0, because `current_time` counts elapsed periods.
 Call these functions on the `parent` of `StroboscopicMap` to obtain the
 corresponding continuous time.
@@ -51,7 +51,8 @@ set_period!(smap::StroboscopicMap, T) = (smap.period = T)
 # Extend interface
 ###########################################################################################
 for f in (:current_state, :initial_state, :current_parameters, :initial_parameters,
-	:dynamic_rule, :set_state!,:successful_step, :(SciMLBase.isinplace), :(StateSpaceSets.dimension))
+	:dynamic_rule, :set_state!, :successful_step, :referrenced_sciml_prob,
+	:(SciMLBase.isinplace), :(StateSpaceSets.dimension))
     @eval $(f)(smap::StroboscopicMap, args...) = $(f)(smap.ds, args...)
 end
 current_time(smap::StroboscopicMap) = smap.t