+++ test/runtests.jl
@@ -14,4 +14,5 @@ testfile(file, testname=defaultname(file)) = @testset "$testname" begin; include
     testfile("poincare.jl")
     testfile("projected.jl")
     testfile("successful_step.jl")
+    testfile("mtk_integration.jl")
 end